+++ src/main/java/com/opentable/db/postgres/junit5/EmbeddedPostgresExtension.java
@@ -14,7 +14,6 @@
 package com.opentable.db.postgres.junit5;
 
 import com.opentable.db.postgres.embedded.DatabasePreparer;
-import org.junit.rules.TestRule;
 
 public final class EmbeddedPostgresExtension {
 
@@ -29,7 +28,7 @@ public final class EmbeddedPostgresExtension {
     }
 
     /**
-     * Returns a {@link TestRule} to create a Postgres cluster, shared amongst all test cases in this JVM.
+     * Returns a {@link PreparedDbExtension} to create a Postgres cluster, shared amongst all test cases in this JVM.
      * The rule contributes Config switches to configure each test case to get its own database.
      * @return PreparedDBExtension
      * @param preparer DatabasePreparer